third_party/WebKit/LayoutTests/fast/events/clientXY-in-zoom-and-scroll-expected.txt

Base
PASS event.clientX is 100
PASS event.clientY is 100
PASS event.pageX is 100
PASS event.pageY is 100

Just zoomed
PASS event.clientX is within 0.00001 of 83.33333
PASS event.clientY is within 0.00001 of 83.33333
PASS event.pageX is within 0.00001 of 83.33333
PASS event.pageY is within 0.00001 of 83.33333

Just scrolled
PASS event.clientX is 100
PASS event.clientY is 100
PASS event.pageX is 150
PASS event.pageY is 150

Zoomed and scrolled
PASS event.clientX is within 0.00001 of 83.33333
PASS event.clientY is within 0.00001 of 83.33333
PASS event.pageX is within 0.00001 of 133.33333
PASS event.pageY is within 0.00001 of 133.33333
PASS successfullyParsed is true

TEST COMPLETE
